# packages/spike.land/docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Spike Land API
  description: |
    API for Spike Land - AI-Powered App Platform with Image Enhancement

    ## Platform Overview
    Spike Land is a platform where users create, fork, and manage AI-powered applications.
    The flagship application, Pixel, provides AI-powered image enhancement.

    ## Authentication
    All protected endpoints require a valid session cookie from NextAuth.js.
    Alternatively, you can use Bearer token authentication for API requests:
    ```
    Authorization: Bearer {session_token}
    ```

    ## Token Economy
    Users have a platform-wide token balance used for:
    - Image enhancement operations (cost depends on resolution tier)
    - Premium features in apps

    Token costs for Pixel image enhancement:
    - TIER_1K: 2 tokens
    - TIER_2K: 5 tokens
    - TIER_4K: 10 tokens

    Tokens regenerate automatically (1 token per 15 minutes, max 100).

    ## Rate Limiting
    Most endpoints are rate-limited per user. Rate limit information is included in response headers:
    - X-RateLimit-Remaining: Requests remaining in current window
    - X-RateLimit-Reset: Unix timestamp when limit resets

    Example rate limits:
    - Image enhancement: 3 requests per minute
    - Image upload: 10 requests per minute
    - Album operations: 10 requests per minute

    ## Error Handling
    The API uses standard HTTP status codes:
    - 200 OK: Successful request
    - 201 Created: Resource created
...

...

    ## Base URL
...
servers:
...
components:
  securitySchemes:
    sessionAuth:
      type: apiKey
      in: cookie
      name: next-auth.session-token
      description: NextAuth.js session cookie (automatically set after login)
    bearerAuth:
      type: http
      scheme: bearer
      description: Bearer token for API requests
  schemas:
...
    Error:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          description: Human-readable error message
          example: "Unauthorized"
        code:
          type: string
          description: Machine-readable error code
          example: "UNAUTHORIZED"
        details:
          type: object
          description: Additional error context
          example: { "field": "email", "reason": "invalid format" }
        title:
          type: string
          description: Short error title
          example: "Authentication Failed"
        suggestion:
          type: string
          description: Suggested action to resolve the error
          example: "Please log in again"
    ValidationError:
      type: object
      required:
        - error
        - field
      properties:
        error:
          type: string
          description: Validation error message
          example: "Invalid input"
        field:
          type: string
          description: Field that failed validation
          example: "email"
        code:
          type: string
          enum:
            - REQUIRED_FIELD
            - INVALID_FORMAT
            - FILE_TOO_LARGE
            - UNSUPPORTED_FORMAT
          description: Validation error code
          example: "INVALID_FORMAT"

    # ========================================
    # Enumerations
    # ========================================
...
  responses:
    Unauthorized:
      description: Authentication required or failed
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error: "Unauthorized"
            code: "UNAUTHORIZED"
            title: "Authentication Failed"
            suggestion: "Please log in again"
    Forbidden:
      description: User lacks permission for this action
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error: "Insufficient permissions"
            code: "FORBIDDEN"
            title: "Access Denied"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error: "Resource not found"
            code: "NOT_FOUND"
            title: "Not Found"
    BadRequest:
      description: Invalid request input
      content:
        application/json:
          schema:
            oneOf:
              - $ref: "#/components/schemas/Error"
              - $ref: "#/components/schemas/ValidationError"
          example:
            error: "Invalid input"
            code: "INVALID_INPUT"
...
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error: "Internal server error"
            code: "INTERNAL_SERVER_ERROR"
            title: "Server Error"
            suggestion: "Please try again later"

  # ========================================
  # Request Parameters
  # ========================================
...
paths:
...
  /api/images/move-to-album:
    post:
      summary: Move images to an album
      tags: [Albums]
      operationId: moveImagesToAlbum
      description: |
        Move one or more images to a target album. Images are added to the album
        and can optionally be removed from a source album. Supports moving up to
        100 images per request. All images and target album must belong to the user.
      security:
        - sessionAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [imageIds, targetAlbumId]
              properties:
                imageIds:
                  type: array
                  items:
                    type: string
                  maxItems: 100
                  minItems: 1
                  description: Array of image IDs to move
                targetAlbumId:
                  type: string
                  description: ID of album to move images to
                removeFromSourceAlbum:
                  type: string
                  nullable: true
                  description: Optional ID of source album to remove images from
      responses:
        "200":
          description: Images moved successfully (partial success allowed)
          content:
            application/json:
              schema:
                type: object
                required: [success, moved, failed]
                properties:
                  success:
                    type: boolean
                  moved:
                    type: integer
                    description: Number of images successfully moved
                  failed:
                    type: integer
                    description: Number of images that failed to move
                  alreadyInAlbum:
                    type: integer
                    description: Number of images already in the album
                  results:
                    type: array
                    items:
                      type: object
                      properties:
                        imageId:
                          type: string
                        success:
                          type: boolean
                        error:
                          type: string
                          nullable: true
        "400":
          description: Invalid input (too many images, missing IDs, or album validation failed)
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
...